rename microcontroller.{beaglebone_black,raspi_23} to {am335x,bcm283x}
These should be named after the chips rather than the boards.
