+++ src/pulseio.py
@@ -3,4 +3,4 @@ from adafruit_blinka.agnostic import board_id
 
 
 if board_id == "raspi_2" or board_id == "raspi_3":
-    from adafruit_blinka.microcontroller.raspi_23.pulseio.PulseIn import PulseIn as PulseIn
+    from adafruit_blinka.microcontroller.bcm283x.pulseio.PulseIn import PulseIn as PulseIn
